Added expandable file-format support & created writer trait to clean GetData.
This commit adds an enum named DataFileFormat so that we can implement more file formats than csv and parquets later.
It also adds a "FileWriter" trait that both CSV writer and Parquet Writer implement so we no longer have to have if statements all the way down in GetData and just focus on
writing the data.