Added setting of TLS certificates
This commit adds TLS certificate setting along with override flags.
A user can set TLS certificates with the set-tls-certificates command to save the paths in config, or use the flags.
I felt like saving the certs themselves in the config was a bad idea, so I instead opted to save the paths and then read the files on run.